Made UI of this feature a bit better, turned off debug code, made a default-setting in the box based on original message
Moves the auto-reply section down to the bottom and gives it some proper writing to indicate how to use it and that BBCode works as well as the fact that people can bypass the message.

Re-implemented the proper check that I was bypassing to test this feature's functionality (the one I commented out with a big warning to myself and almost forgot).

Also went ahead and made a default setting for the box just so that there's some extra info to work with if it's the first time someone uses it.