Dot beside names with original gender colour when using a custom colour (optional)
Puts a dot beside people with custom colours for their names so you can still easily spot the gender while preserving their custom name.

Appears in chat, sidebar, and at the top of private conversations.

Has included option to enable/disable it.

Had to expand the sidebar by 10 pixels to accomodate a name with it that was very long.