Implements an optional gender icon on everyone.
Largely based on the original colour dot code, pretty simple change. Doesn't do the colour thing, adds an icon for the gender to everyone in chat, on the right sidebar, and in searches.

Aimed at a little combo of colourblind assistance and achieving the same goal as the gender dot originally did.