Makes span appear and disappear properly everywhere, also adds an optional colour change for offline users. Needs more testing
Implemented a get into UserView that grabs the specific wanted setting.

Removed some references to a prop that no longer exists as a consequence of doing things this way.

Also added an option to change colours on anyone who logs off to make them consistent with users who are offline in logs. Primarily added for a little consistency with the offline icon that comes with the gender markers.