Updated with a slightly more performance-friendly solution
Does the same as before, changes it almost everywhere instantly.

This route makes it not update in BBCode instantly but a channel change will fix that, it also doesn't update in searches until you do a new search with the setting turned on.

It is also no longer visible in the user's ads tab or shared channels tab. There were a few issues with spacing there and you can still see their gender icon on the main name from there.