Small fix for RecentConversations and CharacterPreview
RecentConversations did not have the new prop integrated, it does now.

Also put the genderClass into CharacterPreview so you see the marker when hovering over someone. This one isn't wholly necessary considering that the gender is written there but the name is coloured there so I figured it wise to put the marker there too.