Options for this now grey out when autoreply is off, also utilises the editor from AdCenter as suggested.
As mentioned, this pulls in the editor from bbcode, utilised in AdCenter. FatCat's suggestion here was far cleverer than my implementation so thank you for that! Also gave me a good place to look to see where to yoink that from with an example of implementation.

It is limited to a fixed value of 10k (5x smaller than the actual message limit in private) on the editor currently rather than importing the maxLength, technically not update resistant but I doubt the max length is changing anytime soon nor that anyone will utilise the full 10k for this ever.

The limit of 10k also leaves it in a position that if those are limited ever, it likely will not dip that far below 50k for private messages which is the only place this should ever fire.

This also makes the custom settings grey out when autoreply is disabled.  This applies both to the requested section and the editor too.

Text describing how this works a little has been updated to reflect the inclusion of an editor and the preview, as well as to include the character limit.